+++ src/components/settings/General.svelte
@@ -1,5 +1,15 @@
-<script>
-    import TextField from "$components/TextField.svelte";
+<script lang="ts">
+    import { getCurrencies } from "$lib/lib.js";
+    import SelectMenu from "$components/SelectMenu.svelte";
+    import { appStore } from "$lib/state.svelte";
+
+    let currencies: string[] = $state([]);
+    $effect(()=>{
+        getCurrencies()
+            .then((c) => {
+                currencies = c;
+            });
+    });
 </script>
 
 <div class="flex items-center justify-between w-full">
@@ -7,5 +17,11 @@
         <p>Currency code</p>
         <p class="text-text-muted text-sm">The ISO currency code</p>
     </div>
-    <TextField value="USD" />
+    <SelectMenu
+        class="w-full max-w-[96px]"
+        defaultValue={appStore.settings.currencyCode}
+        items={currencies}
+        onChange={(c) => appStore.setCurrencyCode(c)}
+        toOption={(item) => {return { label: item,value: item };}}
+    />
 </div>