+++ src/tests/browser/settings.svelte.test.ts
@@ -2,11 +2,23 @@ import { beforeEach, test, expect } from "vitest";
 import { appStore } from "$lib/state.svelte";
 import Sidebar from "$components/Sidebar.svelte";
 import { render } from "vitest-browser-svelte";
+import { mockIPC } from "@tauri-apps/api/mocks";
 
 beforeEach(() => {
     appStore.categories = [];
 });
 
+mockIPC((cmd) => {
+    if (cmd === "fetch_incomes" || cmd === "fetch_expenses") {
+        return [];
+    }
+    if (cmd === "settings") {
+        return { currencyCode: "USD" };
+    }
+    if (cmd === "currencies") {
+        return  ["USD","CAD","ZAR","ZMW","TSH"];
+    }
+});
 test("Open settings panel", async () => {
     const page = render(Sidebar);
     await page.getByLabelText("Open settings").click();
@@ -15,11 +27,11 @@ test("Open settings panel", async () => {
         .toBeInTheDocument();
 });
 
-test("Default to categories section", async () => {
+test("Default to general section", async () => {
     const page = render(Sidebar);
     await page.getByLabelText("Open settings").click();
     const selected = page
-        .getByRole("button", { name: "Categories" })
+        .getByRole("button", { name: "General" })
         .query()
         ?.getAttribute("data-selected");
     expect(selected).toBe("true");
@@ -32,7 +44,7 @@ test("Show categories", async () => {
     ];
     const page = render(Sidebar);
     await page.getByLabelText("Open settings").click();
-    // await page.getByText("Categories").click();
+    await page.getByText("Categories").click();
 
     const items = page.getByRole("listitem").all();
     expect(items).toHaveLength(2);